apps/ml-service/requirements.txt
# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# ML/DL
tensorflow==2.15.0
# o PyTorch: torch==2.1.2 torchvision==0.16.2
scikit-learn==1.4.0
numpy==1.26.3

# Procesamiento de documentos
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10  # OCR si se necesita

# NLP
transformers==4.36.2
sentence-transformers==2.2.2
spacy==3.7.2

# Utilidades
python-dotenv==1.0.0
pandas==2.1.4
joblib==1.3.2
aiofiles==23.2.1
